+++ .eslintrc.js
@@ -31,6 +31,8 @@ module.exports = {
         "object-curly-spacing": ["error", "always", { "objectsInObjects": false }],
         "newline-before-return": "error",
         // "padding-line-between-statements": ["error", { blankLine: "always", prev: "*", next: "return" }]
-        "no-console": ["error"]
+        "no-console": ["error"],
+        "semi": "off",
+        "@typescript-eslint/semi": "error"
     },
 };