+++ src/pipes/validation.pipe.ts
@@ -23,7 +23,7 @@ export class ValidationPipe implements PipeTransform<any> {
 
     private toValidate(metatype: Function): boolean {
         const types: Function[] = [String, Boolean, Number, Array, Object];
-        
+
         return !types.includes(metatype);
     }
 }