+++ src/modules/user/user.service.spec.ts
@@ -4,7 +4,7 @@ import { UserService } from './user.service';
 describe('UserService', () => {
     let service: UserService;
 
-    beforeEach(async () => {
+    beforeEach(async() => {
         const module: TestingModule = await Test.createTestingModule({
             providers: [UserService],
         }).compile();