+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Logger messages to the others modules.
 - [Marek Kurańda](https://github.com/mjkuranda): ESLint rule - `@typescript-eslint/semi`.
+- [Marek Kurańda](https://github.com/mjkuranda): ESLint rule - `no-trailing-spaces`.
 
 ### Changed
 - [Marek Kurańda](https://github.com/mjkuranda): ESLint rule `no-console` disallow all console messages.